requirements.txt
# ==================== Requirements.txt ====================

# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Security
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Development
pytest==7.4.4
httpx==0.26.0
